Support linked xref tables
When a PDF file is updated, the writer MIGHT (according to spec)
add a new xref table, trailer, startxref etc at the end of the file
to make it faster to save large documents. The only requirement is
that the writer includes a pointer to the previous xref table in
its trailer.

This commit will find the last xref table and trailer, and
if the 'Prev' entry exist in the last trailer, read the previous
xref table (and so on).