Restart from scratch
The idea is to treat the underlying COS model as it is,
and build a PDF "layer" above that works directly with
the COS model. Not sure how this will play out, but I guess
we will see...

The idea is to have a `CosDocument` that works with
in-memory storage of objects (for creating PDFs), and a
`CosOpenedDocument` that works with streamed objects AND
an in-memory storage (for loaded PDFs).